moving int to double

+++ PWGJE/Tasks/mctracks_to_aod_softdrop.cxx
@@ -40,11 +40,11 @@ struct MyCustomTask {
   Configurable<double> beta {"beta", 0.2, "beta softdrop parameter"};
   Configurable<int> n {"n", 4, "number of iterations"}; // infinite recursion
 
-  Configurable<int> xmin {"xmin", -1, " xmin "};
-  Configurable<int> xmax {"xmax", 10, " xmax "};
-  Configurable<int> ymin {"ymin", -5, " ymin "};
-  Configurable<int> ymax {"ymax", 10, " ymax "};
-  Configurable<int> nbins {"nbins", 100, " nbins "};
+  Configurable<double> xmin {"xmin", -1, " xmin "};
+  Configurable<double> xmax {"xmax", 10, " xmax "};
+  Configurable<double> ymin {"ymin", -5, " ymin "};
+  Configurable<double> ymax {"ymax", 10, " ymax "};
+  Configurable<double> nbins {"nbins", 100, " nbins "};
 
   void init(o2::framework::InitContext& /*ic*/)
   {
